# co-web/openapi.yaml
openapi: 3.1.0
info:
  title: CO Web API
  description: |
    API da plataforma CO — gestão de conteúdo em grafo.

    Gerado automaticamente a partir de docs/architecture/api-catalog.md.
    Para adicionar ou remover endpoints, atualize o catálogo e rode
    `npm run openapi:gen` em co-web/.

    ## Como adicionar uma rota

    1. Implemente o handler e registre-o no arquivo `*_routes.rs` adequado.
    2. Adicione uma linha na tabela correspondente em `docs/architecture/api-catalog.md`.
    3. Rode `cd co-web && npm run openapi:gen` para regenerar este arquivo.

    ## Envelope opt-in + headers de versão (CO-456 / CO-278-A)

    Toda resposta `/api/v1/*` traz `X-API-Version: 1.0` e
    `X-Co-Server-Version: <versão>`. Envie `X-API-Envelope: 1` (ou
    `Accept: application/vnd.co.v1+json`) para receber o corpo embrulhado no
    schema `Envelope` (`{ data, meta, errors }`). Sem o marcador o corpo é
    idêntico ao formato legado; respostas não-JSON nunca são embrulhadas.
  version: "3.22.0"
  contact:
    name: Arte Longa
    email: yuri@artelonga.com.br
  license:
    name: MIT
servers:
  - url: https://co-artelonga.fly.dev
    description: Production
  - url: http://localhost:8742
    description: Local
...
paths:
...
  /api/v1/ai/query:
    post:
      tags: ["ai"]
      summary: "AI completion query"
      security:
        - sessionCookie: []
        - bearerJWT: []
      responses:
        "200":
          description: OK
  /api/v1/ai/status:
    get:
      tags: ["ai"]
      summary: "AI provider status"
      security:
        - sessionCookie: []
        - bearerJWT: []
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: JWT session cookie (set via /api/v1/auth/verify or password-login)
    apiToken:
      type: http
      scheme: bearer
      description: Long-lived API token (created via POST /api/v1/auth/token)
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token via Authorization header (legacy; prefer sessionCookie)
    githubPAT:
      type: http
      scheme: bearer
      description: GitHub Personal Access Token (gestão administrativa)
    sharedSecret:
      type: apiKey
      in: header
      name: X-Shared-Secret
      description: Shared secret for server-to-server endpoints (e.g. log drains)
...